Extract client secret hasher.
